remove old files, fix duration calculation
